style: eliminate all ast-grep advisory warnings in prod code
- Move Downloader impl from dl/mod.rs to dl/downloader.rs (no-functions-in-mod)
- Move ffmpeg helpers from ffmpeg/mod.rs to ffmpeg/core.rs
- Replace 95+ deep qualified paths with top-level imports across 19 files
- Extract 44+ magic numbers into named constants
- Remove separator comments (// ----) in session.rs, dl/tests.rs, downloader.rs
- Move local `use` to file top in gui/update.rs
- Fix doc backticks for wait_range

+++ tests/tests/kithara_hls/live_stress_real_stream.rs
@@ -900,10 +900,8 @@ async fn live_stress_real_stream_seek_read_cache(
             };
             chunks_read = chunks_read.saturating_add(1);
             if read_idx == 0
-                && let (Some(expected), Some(actual)) = (
-                    expected_variant,
-                    chunk.meta.variant_index.map(|v| v as usize),
-                )
+                && let (Some(expected), Some(actual)) =
+                    (expected_variant, chunk.meta.variant_index.map(|v| v))
             {
                 variant_match_checks = variant_match_checks.saturating_add(1);
                 if expected == actual {
